rag/requirements.txt
cohere==5.15.0
langchain
langchain_community
langchain_cohere
langgraph
tiktoken
jupyter
groq
fastapi
uvicorn
pypdf
faiss-cpu
langchain_groq
PyJWT